rename "ProductList" to "ProductCatalog", move product list
Move product list from "ProductCatalog" in a separate component "ProductList", change styles in "ProductList", add message to empty list
